+++ web/models/page.ex
@@ -110,8 +110,9 @@ defmodule Brando.Page do
     Brando.repo.delete!(record)
   end
   def delete(id) do
-    record = Brando.repo.get_by!(__MODULE__, id: id)
-    delete(record)
+    __MODULE__
+    |> Brando.repo.get_by!(id: id)
+    |> delete
   end
 
   @doc """
@@ -119,8 +120,12 @@ defmodule Brando.Page do
   """
   def order(query) do
     from m in query,
-      order_by: [asc: m.language, asc: m.status,
-                 desc: m.key, desc: m.inserted_at]
+      order_by: [
+        asc: m.language,
+        asc: m.status,
+        desc: m.key,
+        desc: m.inserted_at
+      ]
   end
 
   @doc """